move splitlines
