+++ internal/utils/utils.go
@@ -1,4 +1,4 @@
-package yaml
+package utils
 
 import (
 	"bufio"
